Moved the read and write functions to the main fiat.py script again. This works better. Added a test for the reading of the example ini in test_reader.py.
