+++ hydromt_fiat/workflows/vulnerability.py
@@ -3,17 +3,25 @@ from pathlib import Path
 
 ### TO BE UPDATED ###
 class Vulnerability:
-    def get_susceptibility_function(self):
+    def __init__(self, datadir, config: dict):
+        self.datadir = datadir
+        self.config = config
+
+    def get_vulnerability_function(self):
+        # TODO: @Luis: this function needs to be changed. Now it looks up a damage function
+        # according to country that is chosen but we want the user to choose their
+        # damage functions per category (res, com, etc.). So our function will be
+        # quite different.
         """Return the susceptibility function id and the maximum damage number."""
 
         # Read the global exposure configuration.
         df_config = pd.read_excel(
-            Path(self._DATADIR).joinpath("global_configuration.xlsx"),
+            Path(self.datadir).joinpath("global_configuration.xlsx"),
             sheet_name="Buildings",
         )
 
         # Get the function id.
-        sf_id = df_config.loc[
+        vf_id = df_config.loc[
             df_config["Alpha-3"] == self.config["country"],
             f"Damage_Function_ID_{self.config['hazard_type'].capitalize()}",
         ].values[0]
@@ -24,4 +32,4 @@ class Vulnerability:
             f"Max_Damage_{self.config['hazard_type']. capitalize()}",
         ].values[0]
 
-        return sf_id, max_damage
+        return vf_id, max_damage
