+++ tests/test_vulnerability.py
@@ -1,9 +1,11 @@
 from hydromt_fiat.workflows.vulnerability import Vulnerability
 from pathlib import Path
+import pytest
 
 EXAMPLEDIR = Path().absolute() / "examples"
 
 
+@pytest.mark.skip(reason="Not yet implemented")
 def test_vulnerability_class_initialization():
     # Read model in examples folder.
     Vulnerability()
